Convert null on the wire to an absent document field rather than None
The problem is that if you getAs[String] and it's None, then None
is not a String. The None approach would only work if all other
values were stored as Some(value) instead of just value.
getAs already generates a None if the field is missing, so there's
no need to actually store a None in the map, it works properly
if we just omit the field entirely.

Or another way to put it, we don't want
  getAs[String]("err")
to potentially return Some(None), which it did in
MongoConnectionHandler.