Change CompletableWriteRequest to throw an exception in same cases Java driver would
Java driver CommandResult.java throwOnError throws if !ok or
if the "err" field exists. Previously this code would throw
if the "errmsg" field exists.

Even after looking at the mongo server's C++ for a bit I can't
figure out exactly what "err" vs. "errmsg" is, but ...

One case where CompletableWriteRequest didn't throw, before
this patch, was a duplicate key on insert. This patch
fixes that to complete the future with an exception.