Attempt to rationalize isMaster, connected_?, checkMaster for pooling
For DirectConnection, we can return sensible values, though any
use of these is likely to create race conditions.

For PoolConnection, just pick any single channel with .direct,
and then return whether that channel is connected or master.
If we can think of a better definition later we can use it.

An alternative approach to consider is to remove these
flags from MongoConnection and only support them on
DirectConnection.